Properly implement column relocation, and use it in RotateEditTask
Previously RotateEditTask had hard-coded jank throughout the data-converter, this commit aims to fully separate the two.

To do this a new `EditTask#actOnColumn` api was added, which allows an implementation to output any number of columns at any position.

The cubes outputted by EditTask#actOnCube will attempt to use a column in their position if it exists.